requirements.txt
autoflake==2.2.1
bandit==1.7.5
black==23.10.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
ipaddress==1.0.23
isort==5.12.0
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mypy==1.6.1
mypy-extensions==1.0.0
nornir==3.4.1
nornir-jinja2==0.2.0
nornir-netbox==0.3.0
nornir-utils==0.2.0
packaging==23.2
pathspec==0.11.2
pbr==5.11.1
platformdirs==3.11.0
pydocstyle==6.3.0
pyflakes==3.1.0
Pygments==2.16.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rich==13.6.0
ruamel.yaml==0.18.3
ruamel.yaml.clib==0.2.8
smmap==5.0.1
snowballstemmer==2.2.0
stevedore==5.1.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.5
